Chall 01 antomart
I decided to create a function that switches the letter of a string "as the necklace does" (swithing the last position of the string to the first position). The function compares the 2 strings and "switches" one of them. If there is no diferences between the strings (after swithing positions in one of them) both necklaces are the same. My function also compares the len of both strings.