Create szavarce.py
I have created a dictionary with uppercase letters index, which represent each morse code value. Then checked if there are two arguments to continue, otherwise will get out prompting an error message. Then the program will loop through the argument received (only the arg[1] since arg[0] is the name of the script). If the uppercase char is found in our dictionary, it will add it to a variable to display afterwards. If we find something different from the letters in the dictionary it will get out printing the error message. If we made it to the end, then we print the result stored in "resultado".