chall 03
I know its wrong but i tried! :)
First of all i decided to filter the correct cases. Then I decided to make an iteration count and emulate
snail movements using rmax imax rmin and rmax but that was a clear bad strategy. It sorta works for 2x2 and 3x3 but not for the rest of the cases :(